+++ Changelog.md
@@ -22,6 +22,8 @@ Next Release
 - Fix x265 Slower preset default values (tkozybski)
 - Removal of Menu Styles (Dendraspis, [#589](/../../issues/589))
 - In quality mode main dialog shows count and type of active subtitles (stax76)
+- Main dialog and code editor show script info with F2 key and
+  advanced script info with Ctrl+F2, requires menus to be reset manually 
 
 
 2.1.8.5 Beta (2021-03-09)